+++ src/components/mypage/MyPageMain/MyPageMain.tsx
@@ -1,22 +1,30 @@
 'use client';
 import { MyShelterInfo, MyVolInfo } from '@/api/mypage/mypage';
 import useMyInfo from '@/api/mypage/useMyInfo';
+import useUpdateShelterAlarm from '@/api/mypage/useUpdateShelterAlarm';
+import useUpdateVolInfo from '@/api/mypage/useUpdateVolInfo';
 import {
   ArrowRight,
+  ArrowRightLg,
   MypageChat,
   MypageNoti,
   MypageStar,
   MypageTerms
 } from '@/asset/icons';
-import { Caption2, H3 } from '@/components/common/Typography';
+import ToggleSwitch from '@/components/common/ToggleSwitch/ToggleSwitch';
+import {
+  Body1,
+  Body3,
+  Caption2,
+  Caption3,
+  H3
+} from '@/components/common/Typography';
 import LogoutSection from '@/components/mypage/LogoutSection/LogoutSection';
-import MyStatusItem from '@/components/mypage/MyStatusItem/MyStatusItem';
-import SettingItem from '@/components/mypage/SettingItem/SettingItem';
 import useHeader from '@/hooks/useHeader';
-import uuidv4 from '@/utils/uuidv4';
+import useToast from '@/hooks/useToast';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
-import React from 'react';
+import React, { useCallback } from 'react';
 import * as styles from './MyPageMain.css';
 
 export interface SettingProps {
@@ -35,6 +43,41 @@ export default function MyPageMain({ dangle_role }: { dangle_role: string }) {
   useHeader({ title: '내 정보' });
   const pathname = usePathname();
   const isShelterRole = dangle_role === 'SHELTER';
+
+  const { data: info } = useMyInfo(dangle_role, {
+    enabled: !!dangle_role && dangle_role !== 'NONE'
+  });
+  const { mutateAsync: volunteerAlarm } = useUpdateVolInfo();
+  const { mutateAsync: shelterAlarm } = useUpdateShelterAlarm();
+
+  const toastOn = useToast();
+
+  const handleToggleChange = useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      let payload;
+
+      if (!isShelterInfo(info)) {
+        payload = {
+          nickName: !isShelterInfo(info) ? info?.nickName! : '',
+          phoneNumber: !isShelterInfo(info) ? info?.phoneNumber! : '',
+          alarmEnabled: e.target.checked
+        };
+
+        volunteerAlarm(payload).then(res => {
+          toastOn('카카오톡 알림 설정이 업로드 되었습니다.');
+        });
+      } else {
+        payload = {
+          alarmEnabled: e.target.checked
+        };
+
+        shelterAlarm(payload).then(res => {
+          toastOn('카카오톡 알림 설정이 업로드 되었습니다.');
+        });
+      }
+    },
+    [volunteerAlarm, shelterAlarm, toastOn, info]
+  );
 
   const NOTI: React.ReactNode[] = isShelterRole
     ? [
@@ -61,9 +104,83 @@ export default function MyPageMain({ dangle_role }: { dangle_role: string }) {
     { label: '이용약관', link: '', Svg: MypageTerms }
   ];
 
-  const { data: info } = useMyInfo(dangle_role, {
-    enabled: !!dangle_role && dangle_role !== 'NONE'
-  });
+  if (!info) return null;
+  let myStatus;
+
+  if (isShelterInfo(info)) {
+    const {
+      historyStat: { done, inProgress }
+    } = info;
+
+    myStatus = [
+      { value: '모집 종료', cnt: done },
+      { value: '모집 진행중', cnt: inProgress }
+    ];
+  } else {
+    const {
+      historyStat: { done, waiting, joining }
+    } = info;
+
+    myStatus = [
+      { value: '봉사 이력', cnt: done },
+      { value: '봉사 대기', cnt: waiting },
+      { value: '봉사 신청', cnt: joining }
+    ];
+  }
+
+  function KaKaoAlarm({
+    setting,
+    notification
+  }: {
+    setting: SettingProps;
+    index: number;
+    notification: React.ReactNode[];
+  }) {
+    const { label, Svg } = setting;
+    return (
+      <>
+        <div className={styles.accountBox}>
+          <div className={styles.accountTxt}>
+            <Svg />
+            <Body1>{label}</Body1>
+          </div>
+          <ToggleSwitch
+            name={'alram'}
+            onChange={handleToggleChange}
+            checked={info?.alarmEnabled}
+          />
+        </div>
+
+        <Caption3 color="gray600" className={styles.noti}>
+          {notification as string[]}
+        </Caption3>
+        <div className={styles.divider} />
+      </>
+    );
+  }
+
+  function OtherAlarm({
+    setting,
+    index
+  }: {
+    setting: SettingProps;
+    index: number;
+  }) {
+    const { link, label, Svg } = setting;
+    return (
+      <Link href={link}>
+        <div className={styles.accountBox}>
+          <div className={styles.accountTxt}>
+            <Svg />
+            <Body1>{label}</Body1>
+          </div>
+          <ArrowRightLg />
+        </div>
+
+        {index === 0 || index === 3 ? <div className={styles.divider} /> : null}
+      </Link>
+    );
+  }
 
   return (
     <div className={styles.wrapper}>
@@ -88,21 +205,32 @@ export default function MyPageMain({ dangle_role }: { dangle_role: string }) {
             isShelterRole ? '/shelter/event' : '/volunteer/event'
           }`}
         >
-          <MyStatusItem />
+          <div className={styles.box}>
+            {myStatus.map(({ value, cnt }, index) => (
+              <section style={{ display: 'flex' }} key={index}>
+                <div className={styles.txtGird}>
+                  <H3>{cnt}</H3>
+                  <Body3>{value}</Body3>
+                </div>
+                {index < myStatus.length - 1 && (
+                  <div className={styles.stroke} />
+                )}
+              </section>
+            ))}
+          </div>
         </Link>
       </main>
 
       <div className={styles.settingSection}>
-        {settings.map((setting, index) => (
-          <SettingItem
-            key={uuidv4()}
-            setting={setting}
-            index={index}
-            notification={NOTI}
-            isShelterRole={isShelterRole}
-            dangle_role={dangle_role}
-          />
-        ))}
+        {settings.map((setting, index) => {
+          if (index === 1 && isShelterRole) return null;
+
+          return index === 0 ? (
+            <KaKaoAlarm setting={setting} index={index} notification={NOTI} />
+          ) : (
+            <OtherAlarm setting={setting} index={index} />
+          );
+        })}
       </div>
 
       <LogoutSection />
